DefaultVersionHostname(), RequestID(), Datacenter() need work.
Add TODOs to these functions to either implement support for Go1.11 and
beyond or to officially deprecate them according to the proper
deprecation procedures.
